Cyberdrive_Server/requirements.txt
# CyberDrive Web Server - Requirements

# Core Web
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0

# Serial Communication
pyserial==3.5

# Config & Data
pyyaml>=6.0

# Optional (pour déploiement production)
# gunicorn>=21.2.0
# eventlet>=0.33.3
